Fix showing the latest operations when the given time value spans all audit records.

+++ src/main/java/com/dbw/cli/ShowLatestOperationsOption.java
@@ -7,11 +7,11 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class ShowLatestOperationsOption {
-    private static final int MINUTE_IN_SECONDS = 60;
-    private static final int HOUR_IN_SECONDS = MINUTE_IN_SECONDS * 60;
+    private static final long MINUTE_IN_SECONDS = 60;
+    private static final long HOUR_IN_SECONDS = MINUTE_IN_SECONDS * 60;
 
     private boolean isTime;
-    private short value;
+    private long value;
 
     public boolean isTime() {
         return isTime;
@@ -21,11 +21,11 @@ public class ShowLatestOperationsOption {
         this.isTime = isTime;
     }
 
-    public short getValue() {
+    public long getValue() {
         return value;
     }
 
-    public void setValue(short value) {
+    public void setValue(long value) {
         this.value = value;
     }
 
@@ -38,12 +38,12 @@ public class ShowLatestOperationsOption {
             if (!isValid(input)) {
                 throw new NumberFormatException(ErrorMessages.CLI_INVALID_LATEST_OP);
             }
-            int minutesMultiplier = input.contains("m") ? MINUTE_IN_SECONDS : 1;
-            int hoursMultiplier = input.contains("h") ? HOUR_IN_SECONDS : 1;
-            int extractedNumber = Short.parseShort(input.substring(0, input.length() - 1));
-            int timeInSeconds = extractedNumber * minutesMultiplier * hoursMultiplier;
+            long minutesMultiplier = input.contains("m") ? MINUTE_IN_SECONDS : 1;
+            long hoursMultiplier = input.contains("h") ? HOUR_IN_SECONDS : 1;
+            long extractedNumber = Short.parseShort(input.substring(0, input.length() - 1));
+            long timeInSeconds = extractedNumber * minutesMultiplier * hoursMultiplier;
             opt.setIsTime(true);
-            opt.setValue((short)timeInSeconds);
+            opt.setValue(timeInSeconds);
         }
         return opt;
     }
