+++ src/main/java/com/dbw/db/Orcl.java
@@ -159,7 +159,7 @@ public class Orcl extends Database {
         return maxId;
     }
 
-    public Integer selectLatestAuditRecordId(int seconds) throws SQLException {
+    public Integer selectLatestAuditRecordId(long seconds) throws SQLException {
         return selectSingleIntValue(OrclQueries.SELECT_LATEST_WITH_SECONDS, Common.COLNAME_ID, seconds);
     }
